big chunk of new API, corrected mapper variables from Long to int, added a few mapper functions. Checking new branch due to irregular connection during long train ride.

+++ src/main/java/se/sharpness/hassessolskydd/dao/InstallationDetailsMapper.java
@@ -13,7 +13,7 @@ public interface InstallationDetailsMapper {
     @Select(
             "select * from installation_details where order_id = #{id}"
     )
-    Optional<InstallationDetails> findInstallationDetailsByOrderId(Long id);
+    Optional<InstallationDetails> findInstallationDetailsByOrderId(int id);
 
 
 }
